自宅PCからアップロード 削除の原因がわか～んなぁ。 昔のを見直すしかない

+++ app/src/main/java/com/timeshappy/scoresheet/Item.java
@@ -1,6 +1,6 @@
 package com.timeshappy.scoresheet;
 class Item {
-    private long id;
+    private int id;
     private String count_timer;
     private long date;
     private String number;
@@ -15,7 +15,7 @@ class Item {
 
     public static String numberT;
 
- public Item(long id, String count_timer, long date, String number, int foul, int teamA_foul, int teamB_foul,String teama_name,String teamb_name, int goal_a, int goal_b,String foul_name) {
+ public Item(int id, String count_timer, long date, String number, int foul, int teamA_foul, int teamB_foul,String teama_name,String teamb_name, int goal_a, int goal_b,String foul_name) {
         this.id = id;
         this.count_timer = count_timer;
         this.date = date;
@@ -33,7 +33,7 @@ class Item {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
